+++ packages/mui-material/src/ButtonGroup/ButtonGroup.js
@@ -8,7 +8,7 @@ import getValidReactChildren from '@mui/utils/getValidReactChildren';
 import capitalize from '../utils/capitalize';
 import { styled } from '../zero-styled';
 import memoTheme from '../utils/memoTheme';
-import checkSimplePaletteColorValues from '../utils/checkSimplePaletteColorValues';
+import createSimplePaletteValueFilter from '../utils/createSimplePaletteValueFilter';
 import { useDefaultProps } from '../DefaultPropsProvider';
 import buttonGroupClasses, { getButtonGroupUtilityClass } from './buttonGroupClasses';
 import ButtonGroupContext from './ButtonGroupContext';
@@ -166,7 +166,7 @@ const ButtonGroupRoot = styled('div', {
         },
       },
       ...Object.entries(theme.palette)
-        .filter(([, value]) => value && checkSimplePaletteColorValues(value))
+        .filter(createSimplePaletteValueFilter())
         .flatMap(([color]) => [
           {
             props: { variant: 'text', color },
@@ -230,7 +230,7 @@ const ButtonGroupRoot = styled('div', {
         },
       },
       ...Object.entries(theme.palette)
-        .filter(([, value]) => value && checkSimplePaletteColorValues(value, ['dark']))
+        .filter(createSimplePaletteValueFilter(['dark']))
         .map(([color]) => ({
           props: { variant: 'contained', color },
           style: {
