+++ packages/mui-material/src/FilledInput/FilledInput.js
@@ -8,7 +8,7 @@ import InputBase from '../InputBase';
 import rootShouldForwardProp from '../styles/rootShouldForwardProp';
 import { styled } from '../zero-styled';
 import memoTheme from '../utils/memoTheme';
-import checkSimplePaletteColorValues from '../utils/checkSimplePaletteColorValues';
+import createSimplePaletteValueFilter from '../utils/createSimplePaletteValueFilter';
 import { useDefaultProps } from '../DefaultPropsProvider';
 import filledInputClasses, { getFilledInputUtilityClass } from './filledInputClasses';
 import {
@@ -138,7 +138,7 @@ const FilledInputRoot = styled(InputBaseRoot, {
           },
         },
         ...Object.entries(theme.palette)
-          .filter(([, value]) => value && checkSimplePaletteColorValues(value)) // check all the used fields in the style below
+          .filter(createSimplePaletteValueFilter()) // check all the used fields in the style below
           .map(([color]) => ({
             props: {
               disableUnderline: false,
