+++ packages/mui-material/src/OutlinedInput/OutlinedInput.js
@@ -9,7 +9,7 @@ import formControlState from '../FormControl/formControlState';
 import rootShouldForwardProp from '../styles/rootShouldForwardProp';
 import { styled } from '../zero-styled';
 import memoTheme from '../utils/memoTheme';
-import checkSimplePaletteColorValues from '../utils/checkSimplePaletteColorValues';
+import createSimplePaletteValueFilter from '../utils/createSimplePaletteValueFilter';
 import { useDefaultProps } from '../DefaultPropsProvider';
 import outlinedInputClasses, { getOutlinedInputUtilityClass } from './outlinedInputClasses';
 import InputBase, {
@@ -56,7 +56,7 @@ const OutlinedInputRoot = styled(InputBaseRoot, {
       },
       variants: [
         ...Object.entries(theme.palette)
-          .filter(([, value]) => value && checkSimplePaletteColorValues(value))
+          .filter(createSimplePaletteValueFilter())
           .map(([color]) => ({
             props: { color },
             style: {
